+++ crates/bindgen/Cargo.toml
@@ -14,4 +14,5 @@ name = "qiskit_bindgen"
 
 [dependencies]
 anyhow.workspace = true
-cbindgen.workspace = true
+cbindgen  = { workspace = true, features = ["unstable_ir"] }
+hashbrown.workspace = true
