+++ crates/pyext/Cargo.toml
@@ -24,8 +24,7 @@ cache_pygates = ["qiskit-circuit/cache_pygates", "qiskit-accelerate/cache_pygate
 
 [build-dependencies]
 anyhow.workspace = true
-cbindgen = { workspace = true, features = ["unstable_ir"] }
-hashbrown.workspace = true
+cbindgen.workspace = true
 qiskit-bindgen.workspace = true
 qiskit-cext-vtable = { workspace = true, features = ["python_binding"] }
